Create portScanner
This post is just to provide a sample code to generate a Port Scanner. This Port Scanner will work for both the Web Applications and remote Host. This tool has been created to provide the basic functionality of a Port Scanner. The general concept of Sockets had been used to provide the functionality. Port Scanner is built on Python 3 and uses some extra libraries such as socket and pyfiglet (for a fancy banner).

Please find the below source code for the Port Scanner attached here.